Modularización de las funciones de escritura y lectura de archivos
