+++ AeroTaxi/src/com/company/JFrames/cuestionario.java
@@ -4,10 +4,10 @@ import com.company.Airplane.Airplane;
 import com.company.Airplane.Planes.Bronze;
 import com.company.Airplane.Planes.Gold;
 import com.company.Airplane.Planes.Silver;
-import com.company.Cityy.City;
+import com.company.City.City;
 import com.company.CompanyAdmin.Company;
 import com.company.Flight.Flight;
-import com.company.Questionaryy.Questionary;
+import com.company.Questionary.Questionary;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.toedter.calendar.JCalendar;
 import com.toedter.calendar.JDateChooser;
