+++ AeroTaxi/src/com/company/JFrames/verifyUser.java
@@ -1,7 +1,7 @@
 package com.company.JFrames;
 
 import com.company.CompanyAdmin.Company;
-import com.company.Userr.User;
+import com.company.User.User;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import javax.swing.*;
